flask-backend/requirements.txt
astroid==2.3.3
autopep8==1.4.4
beautifulsoup4==4.8.1
botocore==1.14.9
bs4==0.0.1
cement==2.8.2
certifi==2019.11.28
chardet==3.0.4
Click==7.0
colorama==0.3.9
dnspython==1.16.0
docutils==0.15.2
Flask==1.1.1
flask-cors==3.0.8
future==0.16.0
gunicorn==20.0.4
idna==2.7
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.3
jmespath==0.9.4
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
pathspec==0.5.9
pycodestyle==2.5.0
pylint==2.4.4
pymongo==3.10.0
python-aqi==0.6.1
python-dateutil==2.8.0
pytz==2019.3
PyYAML==5.2
requests==2.20.1
semantic-version==2.5.0
six==1.11.0
soupsieve==1.9.5
sqlparse==0.3.0
termcolor==1.1.0
urllib3==1.24.3
virtualenv==16.7.9
wcwidth==0.1.8
Werkzeug==0.16.0
wrapt==1.11.2
